[broken] Build things to wire into solved hopes
